requirements.txt
pygame-ce>=2.2.1